correction erreur conversion
git-svn-id: svn+ssh://cvs-etis.ensea.fr/svn/SimulateurV1/pandora/trunk@84 fd298be7-36cc-4ec4-a0e8-09baf7357fce

+++ save/convert_matlab.py
@@ -53,7 +53,7 @@ nbriteration=donnees.shape[0]/ligne
 #filtre les warning qui viennent ensuite
 warnings.filterwarnings("ignore")
 #Lis le fichier excepte les deux premieres ligne, en prenant en compte les # donc les time, vire toute les ligne qui ne sont pas de la taille de la premiere traitee (cest a dire time qui est de taille 2)
-data_temps= np.genfromtxt(fichier,skiprows=2,comments='a',invalid_raise=False,dtype=int)
+data_temps= np.genfromtxt(fichier,skiprows=2,comments='a',invalid_raise=False,dtype=float)
 #remet les warnings apres cette fonction
 warnings.resetwarnings()
 #ne recupere que la seconde colonne precedente (les temps)
